requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
sentence-transformers
pypdf
faiss-cpu
python-dotenv